Expose channel adapter as a bean
WebSocketMessageDrivenChannelAdapter needs to be exposed as a bean so
that it's lifecycle is driven. The creation of the request handler
remains in a factory bean, but the creation of the channel adapter
is now done 'directly' in the namespace parser, with a reference to the
channel adapter being passed into the factory bean for the request
handler.